FP2: enable tmpfs's ACL support in config.
This is used by Systemd's uaccess helper on Ubuntu to grant permission
to some device node based on current active user ("seat"). In this
particular case, it's the video nodes for hardware encoder/decoder, so
this fix video playback on Ubuntu touch.

Also, make config available via /proc/config.gz.

Note that this is the result of "make menuconfig". The previous config
seems not to be generated by Linux's Makefile, so, the diff will be
pretty large.